Fix #8 by adding a new changelog file to the src tree.
With the change to the Github issue tracker it becomes required to add a
different way of handling the changelog of the plugin. This had been
done by adding the changelog file directly to the src. In the future
we'll now only reference/copy the reflected parts into the wiki making
it easier to maintain and find the changes.

Signed-off-by: Andreas Pieber <anpieber@gmail.com>